rename & clean
