+++ packages/core/src/base/skill.ts
@@ -64,13 +64,13 @@ import {
 } from "../utils";
 import type { IDetailLogger } from "../log";
 import type { CustomEvent } from "./custom_event";
-import { getRaw, NoReactiveSymbol } from "../runtime/context/reactive";
-import type { PlainCharacterState } from "../runtime/context/utils";
+import { getRaw, NoReactiveSymbol } from "../runtime/reactive/reactive";
+import type { PlainCharacterState } from "../runtime/reactive/utils";
 import type { AppliableDamageType } from "../data/type";
 import type { MoveEntityM, RemoveEntityM } from "./mutation";
 import type { LunarReaction } from "@gi-tcg/typings";
 import type { DamageOption, ReadonlyEventList } from "../mutator";
-import type { SkillContext } from "../runtime/context/skill";
+import type { SkillContext } from "../runtime/skill_context";
 
 export interface SkillDefinitionBase<Arg> {
   readonly type: "skill";
