+++ packages/core/src/reaction.ts
@@ -14,7 +14,7 @@ import type {
   SummonHandle,
 } from "./data/type";
 import { $ } from "./query";
-import { SkillContext, type TypedSkillContext } from "./runtime/context/skill";
+import { SkillContext, type TypedSkillContext } from "./runtime/skill_context";
 
 export const CALLED_FROM_REACTION: unique symbol = Symbol();
 
@@ -41,7 +41,6 @@ type ReactionContextMeta = {
   eventArgType: ReactionDescriptionEventArg;
   callerType: never;
   associatedExtension: never;
-  shortcutReceiver: unknown;
   gtsSnippets: {};
 };
 type ReactionAction = (context: TypedSkillContext<ReactionContextMeta>) => void;
