+++ packages/core/src/runtime/reactive/attachment.ts
@@ -13,32 +13,40 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import type { EntityState, EntityVariables } from "../../base/state";
+import type {
+  AttachmentState,
+  EntityState,
+  EntityVariables,
+} from "../../base/state";
 import { GiTcgDataError } from "../../error";
 import { type EntityArea, type EntityDefinition } from "../../base/entity";
-import { diceCostSizeOfCard, getEntityById } from "./utils";
-import type { ContextMetaBase, SkillContext } from "./skill";
+import { getEntityById } from "./utils";
+import type { ContextMetaBase, SkillContext } from "../skill_context";
 import {
   LatestStateSymbol,
   RawStateSymbol,
   ReactiveStateBase,
   ReactiveStateSymbol,
 } from "./reactive_base";
-import type { AttachmentHandle } from "../../data/type";
+import type { AttachmentDefinition } from "../../base/attachment";
+import type { RxEntityState } from "./reactive";
 
-class ReadonlyEntity<Meta extends ContextMetaBase> extends ReactiveStateBase {
-  override get [ReactiveStateSymbol](): "entity" {
-    return "entity";
+class ReadonlyAttachment<
+  Meta extends ContextMetaBase,
+> extends ReactiveStateBase {
+  override get [ReactiveStateSymbol](): "attachment" {
+    return "attachment";
   }
-  declare [RawStateSymbol]: EntityState;
-  override get [LatestStateSymbol](): EntityState {
+  declare [RawStateSymbol]: AttachmentState;
+  override get [LatestStateSymbol](): AttachmentState {
     const state = getEntityById(
       this.skillContext.rawState,
       this.id,
-    ) as EntityState;
+    ) as AttachmentState;
     return state;
   }
 
+  // protected _area: EntityArea | undefined;
   constructor(
     protected readonly skillContext: SkillContext<Meta>,
     public readonly id: number,
@@ -46,10 +54,10 @@ class ReadonlyEntity<Meta extends ContextMetaBase> extends ReactiveStateBase {
     super();
   }
 
-  protected get state(): EntityState {
+  protected get state(): AttachmentState {
     return this[LatestStateSymbol];
   }
-  get definition(): EntityDefinition {
+  get definition(): AttachmentDefinition {
     return this.state.definition;
   }
   get area(): EntityArea {
@@ -67,29 +75,19 @@ class ReadonlyEntity<Meta extends ContextMetaBase> extends ReactiveStateBase {
     return this.state.variables[name];
   }
 
-  /** 当前元素骰费用 */
-  diceCost() {
-    return diceCostSizeOfCard(this.skillContext.rawState, this.latest());
-  }
-
-  withAttachment(id: AttachmentHandle) {
-    return this.state.attachments.some(att => att.definition.id === id);
-  }
-  empowered() {
-    // Empowerment: 206
-    return this.withAttachment(206 as AttachmentHandle);
-  }
-
-
-  get master() {
-    if (this.area.type !== "characters") {
-      throw new GiTcgDataError("master expect a character area");
+  get master(): RxEntityState<Meta, "eventCard" | "support" | "equipment"> {
+    if (this.area.type !== "hands" && this.area.type !== "pile") {
+      throw new GiTcgDataError("master expect a hands/pile area");
     }
-    return this.skillContext.get<"character">(this.area.characterId);
+    return this.skillContext.get<"eventCard" | "support" | "equipment">(
+      this.area.cardId,
+    );
   }
 }
 
-export class Entity<Meta extends ContextMetaBase> extends ReadonlyEntity<Meta> {
+export class Attachment<
+  Meta extends ContextMetaBase,
+> extends ReadonlyAttachment<Meta> {
   setVariable(prop: string, value: number) {
     this.skillContext.setVariable(prop, value, this.state);
   }
@@ -99,9 +97,6 @@ export class Entity<Meta extends ContextMetaBase> extends ReadonlyEntity<Meta> {
   addVariableWithMax(prop: string, value: number, maxLimit: number) {
     this.skillContext.addVariableWithMax(prop, value, maxLimit, this.state);
   }
-  consumeUsage(count = 1) {
-    this.skillContext.consumeUsage(count, this.state);
-  }
   resetUsagePerRound() {
     this.skillContext.mutate({
       type: "resetVariables",
@@ -109,10 +104,12 @@ export class Entity<Meta extends ContextMetaBase> extends ReadonlyEntity<Meta> {
       state: this.state,
     });
   }
-  dispose() {
-    this.skillContext.dispose(this.state);
+  dispose(): never {
+    throw new GiTcgDataError(
+      "Attachment can not be disposed directly, for now",
+    );
   }
 }
 
-export type TypedEntity<Meta extends ContextMetaBase> =
-  Meta["readonly"] extends true ? ReadonlyEntity<Meta> : Entity<Meta>;
+export type TypedAttachment<Meta extends ContextMetaBase> =
+  Meta["readonly"] extends true ? ReadonlyAttachment<Meta> : Attachment<Meta>;
