+++ packages/core/src/runtime/skill.ts
@@ -45,11 +45,9 @@ import {
   type ContextMetaBase,
   SkillContext,
   type TypedSkillContext,
-} from "./context/skill";
+} from "./skill_context";
 import type { ExEntityType, ExtensionHandle, SkillHandle } from "../data/type";
-import {
-  type EntityArea,
-} from "../base/entity";
+import { type EntityArea } from "../base/entity";
 import {
   getActiveCharacterIndex,
   getEntityArea,
@@ -58,26 +56,21 @@ import {
   normalizeCost,
 } from "../utils";
 import { GiTcgDataError } from "../error";
-import type { ApplyReactive } from "./context/reactive";
+import type { ApplyReactive } from "./reactive/reactive";
 
 export type InitiativeSkillTargetKind = readonly (
-  | "character"
-  | "summon"
-  | "support"
+  "character" | "summon" | "support"
 )[];
 
-export type SkillContextMeta = Omit<
-  ContextMetaBase,
-  "readonly" | "shortcutReceiver"
->;
-export type ReadonlyMetaOf<BM extends SkillContextMeta> = {
-  [K in keyof SkillContextMeta]: BM[K];
-} & { readonly: true; shortcutReceiver: unknown };
-export type WritableMetaOf<BM extends SkillContextMeta> = {
-  [K in keyof SkillContextMeta]: BM[K];
-} & { readonly: false; shortcutReceiver: unknown };
+export type RwContextMeta = Omit<ContextMetaBase, "readonly">;
+export type ReadonlyMetaOf<BM extends RwContextMeta> = {
+  [K in keyof RwContextMeta]: BM[K];
+} & { readonly: true };
+export type WritableMetaOf<BM extends RwContextMeta> = {
+  [K in keyof RwContextMeta]: BM[K];
+} & { readonly: false };
 
-export type SkillOperation<Meta extends SkillContextMeta> = (
+export type SkillOperation<Meta extends RwContextMeta> = (
   c: TypedSkillContext<WritableMetaOf<Meta>>,
   e: ApplyReactive<
     WritableMetaOf<Meta>,
@@ -85,7 +78,7 @@ export type SkillOperation<Meta extends SkillContextMeta> = (
   >,
 ) => void;
 
-export type SkillOperationFilter<Meta extends SkillContextMeta> = (
+export type SkillOperationFilter<Meta extends RwContextMeta> = (
   c: TypedSkillContext<ReadonlyMetaOf<Meta>>,
   e: ApplyReactive<
     ReadonlyMetaOf<Meta>,
@@ -93,7 +86,7 @@ export type SkillOperationFilter<Meta extends SkillContextMeta> = (
   >,
 ) => unknown;
 
-type SkillProjection<Projected, Meta extends SkillContextMeta> = (
+type SkillProjection<Projected, Meta extends RwContextMeta> = (
   c: TypedSkillContext<ReadonlyMetaOf<Meta>>,
   e: ApplyReactive<
     ReadonlyMetaOf<Meta>,
@@ -633,7 +626,7 @@ export function wrapSkillInfoWithExt(
   return {
     ...skillInfo,
     associatedExtensionId,
-    gtsSnippets: new Map()
+    gtsSnippets: new Map(),
   };
 }
 
@@ -650,7 +643,7 @@ function generateTargetList(
     return [[]];
   }
   const [first, ...rest] = getTarget;
-  const ctx = new SkillContext<ReadonlyMetaOf<SkillContextMeta>>(
+  const ctx = new SkillContext<ReadonlyMetaOf<RwContextMeta>>(
     state,
     wrapSkillInfoWithExt(skillInfo, associatedExtensionId),
     {
